refactor: move atproto and format enums to _internal
- move src/backend/atproto/ to src/backend/_internal/atproto/
- move audio.py and image.py format enums from models/ to _internal/
- update all imports accordingly including tests
- remove AudioFormat from models/__init__.py exports

these are internal utilities, not database models. organizing them under
_internal/ makes the codebase structure clearer.
