+++ src/backend/_internal/atproto/__init__.py
@@ -1,7 +1,7 @@
 """ATProto integration for relay."""
 
-from backend.atproto.profile import fetch_user_avatar, fetch_user_profile
-from backend.atproto.records import (
+from backend._internal.atproto.profile import fetch_user_avatar, fetch_user_profile
+from backend._internal.atproto.records import (
     create_like_record,
     create_track_record,
     delete_record_by_uri,
