+++ src/backend/_internal/image.py
