+++ src/backend/models/__init__.py
@@ -1,7 +1,6 @@
 """database models."""
 
 from backend.models.artist import Artist
-from backend.models.audio import AudioFormat
 from backend.models.database import Base
 from backend.models.exchange_token import ExchangeToken
 from backend.models.oauth_state import OAuthStateModel
@@ -14,7 +13,6 @@ from backend.utilities.database import db_session, get_db, init_db
 
 __all__ = [
     "Artist",
-    "AudioFormat",
     "Base",
     "ExchangeToken",
     "OAuthStateModel",
