+++ src/backend/utilities/aggregations.py
@@ -27,4 +27,4 @@ async def get_like_counts(db: AsyncSession, track_ids: list[int]) -> dict[int, i
     )
 
     result = await db.execute(stmt)
-    return dict(result.all())
+    return dict(result.all())  # type: ignore
